requirements.txt
requests==2.22.0
numpy==1.17.4
pandas==1.0.3
pydeck==0.4.0b2
streamlit==0.63.0
Pillow==7.2.0
